i18n domains validation as a util
